backend/requirements.txt
# FixMyCondo Backend - FastAPI + Redis + SQLite
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis for Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File Handling
aiofiles==23.2.1
Pillow==10.2.0

# CORS
starlette==0.35.1

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.0
